+++ src/store/mix_refresh/index.d.ts
@@ -1,8 +1,6 @@
-import type { Subscriber } from 'svelte/store'
+import type { Readable, Subscriber, Writable } from 'svelte/store'
 import type { refresh_mixin_T } from '../_types'
-import type { Readable_ } from '../readable_'
-import type { Writable_ } from '../writable_'
 export declare function mix_refresh<store_T extends unknown = unknown>(
-	readable:Readable_<store_T>,
-	set?:typeof readable extends Writable_<store_T> ? undefined : Subscriber<store_T>
+	readable:Readable<store_T>,
+	set?:typeof readable extends Writable<store_T> ? undefined : Subscriber<store_T>
 ):refresh_mixin_T<store_T>