+++ src/store/writable_fn_/index.d.ts
@@ -1,3 +1,3 @@
 import type { Writable } from 'svelte/store'
 import type { Writable_ } from '../writable_'
-export declare function writable_fn_<T>(atom:Writable<T>):Writable_<T>
+export declare function writable_fn_<T>(writable:Writable<T>):Writable_<T>