+++ src/store/writable_fn_/index.js
@@ -1,21 +1,21 @@
 import { get } from 'svelte/store'
-export function writable_fn_(atom) {
+export function writable_fn_(writable) {
 	function fn(_) {
-		return arguments.length ? atom.set(_) : get(atom)
+		return arguments.length ? writable.set(_) : get(writable)
 	}
 	return new Proxy(fn, {
 		get(target, prop, receiver) {
 			if (prop === '_' || prop === '$') {
-				return get(atom)
+				return get(writable)
 			}
-			return Reflect.get(atom, prop, atom)
+			return Reflect.get(writable, prop, writable)
 		},
 		set(target, prop, val, receiver) {
 			if (prop === '_' || prop === '$') {
-				atom.set(val)
+				writable.set(val)
 				return true
 			}
-			return Reflect.set(atom, prop, val, atom)
+			return Reflect.set(writable, prop, val, writable)
 		}
 	})
 }